refactor(ops): group file operations under arlmet.ops
Move concat, subset, and sampling into an arlmet/ops/ subpackage, and rename
sampling.py -> sample.py for consistency with concat/subset. The public
top-level API is unchanged: arlmet.concat, arlmet.concat_by_time,
arlmet.extract_subset, and arlmet.sample_points still import from arlmet
directly. Only code importing the submodules by path is affected
(arlmet.subset -> arlmet.ops.subset, arlmet.sampling -> arlmet.ops.sample).

file.py now imports the sampling op lazily inside File.sample_points (matching
File.extract_subset), so file no longer imports ops at module load. This keeps
the file <-> ops dependency one-directional and avoids an import cycle (ops sits
on top of file).
