+++ src/arlmet/ops/sample.py
