+++ src/arlmet/ops/subset.py
